refractored inputs
