+++ src/components/form-input/form-input.component.js
@@ -1,14 +1,23 @@
 import React from 'react';
-import classes from './form-input.module.scss'
+import { useField } from 'formik';
+import classes from './form-input.module.scss';
 
-const FormInput = ({ handleChange, label,  ...otherProps }) => (
-    <div className={classes.group}>
-        <input className={classes.formInput} onChange={ handleChange } {...otherProps} />
-        {
-            label ? (<label className={`${otherProps.value.length ? `${classes.shrink}` : ''} ${classes.formInputLabel}`}>
-                { label }</label>) : null
-        }
-    </div>
-)
+const TextInput = ({ label, ...props }) => {
+    const [field, meta] = useField(props);
+    return (
+        <div className={classes.group}>
+            <input className={classes.formInput} {...field} {...props} />
+            {meta.touched && meta.error ? (
+                <div className={classes.error}>{meta.error}</div>
+            ) : null}
 
-export default FormInput;
+            { label ? (<label
+                className={`${field.value.length ? `${classes.shrink}` : ''} 
+                ${classes.formInputLabel}`}>{label}</label>)
+                : null}
+        </div>
+    );
+
+};
+
+export default TextInput;
