+++ src/components/form-input/form-input.module.scss
@@ -45,3 +45,8 @@
     }
   }
 }
+
+.error {
+  color: $errorColor;
+  font-size: 14px;
+}
