--- src/components/text-input/text-input.module.scss
@@ -1,52 +1,0 @@
-@import '../../baseScss/global/variables';
-@import '../../baseScss/global/mixins';
-
-.group {
-  position: relative;
-  margin: 45px 0;
-
-  .formInput {
-    background: #ffffff none;
-    color: $formInputSub;
-    font-size: 18px;
-    padding: 10px 10px 10px 5px;
-    display: block;
-    width: 100%;
-    border: none;
-    border-radius: 0;
-    border-bottom: 1px solid $formInputSub;
-    margin: 25px 0;
-
-    &:focus {
-      outline: none;
-    }
-
-    &:focus ~ .form-input-label {
-      @include shrinkLabel();
-    }
-  }
-
-  input[type='password'] {
-    letter-spacing: 0.3em;
-  }
-
-  .formInputLabel {
-    color: $formInputSub;
-    font-size: 16px;
-    font-weight: normal;
-    position: absolute;
-    pointer-events: none;
-    left: 5px;
-    top: 10px;
-    transition: 300ms ease all;
-
-    &.shrink {
-      @include shrinkLabel();
-    }
-  }
-}
-
-.error {
-  color: $errorColor;
-  font-size: 14px;
-}
