+++ src/pages/login/login.page.js
@@ -5,7 +5,7 @@ import { auth, signInWithGoogle } from "../../firebase/firebase.utils";
 import { Link } from "react-router-dom";
 import { Formik, Form } from 'formik';
 import * as Yup from 'yup';
-import TextInput from "../../components/text-input/text-input.component";
+import TextInput from "../../components/form-input/form-input.component";
 import ErrorComponent from "../../components/error-component/error-component";
 
 const LoginPage = () => {
